moved config files
