+++ SparseGrids/tsgCudaLinearAlgebra.hpp
@@ -31,7 +31,7 @@
 #ifndef __TASMANIAN_SPARSE_GRID_CUDA_LINEAR_ALGEBRA_HPP
 #define __TASMANIAN_SPARSE_GRID_CUDA_LINEAR_ALGEBRA_HPP
 
-#include "tasmanianConfig.hpp"
+#include "TasmanianConfig.hpp"
 
 #ifdef TASMANIAN_CUDA
