+++ SparseGrids/tsgCudaMacros.hpp
@@ -31,7 +31,7 @@
 #ifndef __TASMANIAN_SPARSE_GRID_CUDA_MACROS_HPP
 #define __TASMANIAN_SPARSE_GRID_CUDA_MACROS_HPP
 
-#include "tasmanianConfig.hpp"
+#include "TasmanianConfig.hpp"
 
 #include <iostream>
